add validation on backend

+++ backend/parking-backend/src/main/java/ru/app/parking_backend/repository/CarRepository.java
@@ -43,32 +43,30 @@ public class CarRepository {
         return jdbc.query(sql, dtoMapper);
     }
 
-    // ищет машину по id
+    // машина по id
     public Optional<Car> findById(Integer id) {
         List<Car> cars = jdbc.query("SELECT * FROM cars WHERE id = ?", rowMapper, id);
         return cars.stream().findFirst();
     }
 
-    // ищет машину по номеру используя частичное совпадение
+    // поиск машины по номеру используя совпадение ILIKE
     public List<CarDto> searchByNumber(String numberCar) {
         String sql = "SELECT c.*, cl.full_name as client_full_name FROM cars c LEFT JOIN clients cl ON c.client_id = cl.id WHERE c.number_car ILIKE ?";
         return jdbc.query(sql, dtoMapper, "%" + numberCar + "%");
     }
 
-    // поиск машины с владельцем по номеру
-// сохраняет новую или обновляет существующую машину
+    // сохранение/обновление машины
     public Car save(Car car) {
         if (car.id() == null) {
             // keyholder нужен чтобы после вставки получить id который база сама сгенерировала
-            // мы используем preparedstatement с указанием что нам нужны ключи
-            // база возвращает ключи и мы сохраняем их в keyholder
+            // благодаря preparedStatement база возвращает ключи и мы сохраняем их в keyholder
             // потом достаем id и ставим его объекту машины
             KeyHolder keyHolder = new GeneratedKeyHolder();
             jdbc.update(connection -> {
                 PreparedStatement ps = connection.prepareStatement("INSERT INTO cars (number_car, client_id) VALUES (?, ?)", Statement.RETURN_GENERATED_KEYS);
                 ps.setString(1, car.numberCar());
                 if (car.clientId() != null) {
-                    ps.setLong(2, car.clientId());
+                    ps.setInt(2, car.clientId());
                 } else {
                     ps.setNull(2, Types.INTEGER);
                 }
@@ -83,7 +81,6 @@ public class CarRepository {
             return car;
         }
     }
-
 
     public void delete(Integer id) {
         jdbc.update("DELETE FROM cars WHERE id = ?", id);
